backend/requirements.txt
blinker==1.6.2
click==8.1.3
dnspython==2.3.0
Flask==2.3.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.5.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mongomock==4.1.2
mypy==1.3.0
mypy-extensions==1.0.0
packaging==23.1
pluggy==1.0.0
PyJWT==2.7.0
pymongo==4.3.3
pytest==7.3.2
sentinels==1.0.0
six==1.16.0
typing_extensions==4.6.3
Werkzeug==2.3.6
python-dateutil==2.8.2